+++ cli/collection_create.go
@@ -17,7 +17,9 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/sourcenetwork/defradb/client"
+	"github.com/sourcenetwork/defradb/datastore"
 	"github.com/sourcenetwork/defradb/internal/db"
+	"github.com/sourcenetwork/defradb/internal/encryption"
 )
 
 func MakeCollectionCreateCommand() *cobra.Command {
@@ -111,4 +113,17 @@ Example: create from stdin:
 		"Comma-separated list of fields to encrypt")
 	cmd.Flags().StringVarP(&file, "file", "f", "", "File containing document(s)")
 	return cmd
+}
+
+// setContextDocEncryption sets doc encryption for the current command context.
+func setContextDocEncryption(cmd *cobra.Command, shouldEncrypt bool, encryptFields []string, txn datastore.Txn) {
+	if !shouldEncrypt && len(encryptFields) == 0 {
+		return
+	}
+	ctx := cmd.Context()
+	if txn != nil {
+		ctx = encryption.ContextWithStore(ctx, txn)
+	}
+	ctx = encryption.SetContextConfigFromParams(ctx, shouldEncrypt, encryptFields)
+	cmd.SetContext(ctx)
 }
